Added Methods to Associate Remote Addresses with Proxies
This makes it so that even if an address won't be needed for a connection to a proxy we can still know that it is available. For example, in the case of TCP both peer's host a listener port, but only one of the peers will connect as a "client". 
- Might need to go back and improve storage usage because the address is also stored in the RegisteredEndpoint object.